fix: make contributor tests server-safe
