+++ src/lib/submissions.ts
@@ -4,13 +4,28 @@
 // MUST come from a verified caller (Clerk session for web, OAuth bearer for
 // CLI) — never from request body — so this module accepts them as `principal`.
 
+import "server-only";
+
 import { eq } from "drizzle-orm";
 import { Resend } from "resend";
 
 import { db, schema } from "@/lib/db/client";
 import { renderNewSubmissionEmail } from "@/lib/email-templates/new-submission";
+import {
+  type SubmissionInput,
+  type SubmissionResult,
+  slugify,
+} from "@/lib/submissions-validation";
 import { getPreferredLocaleForUser } from "@/lib/user-locale";
-import { isAllowedAssetUrl } from "@/lib/url-allowlist";
+
+export {
+  MIN_SPRITE_DIM,
+  REQUIRED_FIELDS,
+  type SubmissionInput,
+  type SubmissionResult,
+  slugify,
+  validateSubmission,
+} from "@/lib/submissions-validation";
 
 export type SubmissionPrincipal = {
   userId: string;
@@ -22,89 +37,6 @@ export type SubmissionPrincipal = {
   /** Pre-computed external profile URL (X or GitHub) when available. */
   url?: string | null;
 };
-
-export type SubmissionInput = {
-  zipUrl: string;
-  spritesheetUrl: string;
-  petJsonUrl: string;
-  displayName: string;
-  description: string;
-  petId: string;
-  spritesheetWidth: number;
-  spritesheetHeight: number;
-};
-
-export type SubmissionResult =
-  | { ok: true; id: string; slug: string }
-  | {
-      ok: false;
-      status: number;
-      error: string;
-      message?: string;
-      field?: string;
-      got?: unknown;
-    };
-
-export const REQUIRED_FIELDS: ReadonlyArray<keyof SubmissionInput> = [
-  "zipUrl",
-  "spritesheetUrl",
-  "petJsonUrl",
-  "displayName",
-  "description",
-  "petId",
-  "spritesheetWidth",
-  "spritesheetHeight",
-] as const;
-
-export const MIN_SPRITE_DIM = 256;
-
-export function validateSubmission(
-  body: Partial<SubmissionInput>,
-): SubmissionResult | null {
-  for (const field of REQUIRED_FIELDS) {
-    if (!body[field]) {
-      return {
-        ok: false,
-        status: 400,
-        error: "missing_field",
-        field,
-      };
-    }
-  }
-  if (
-    !body.spritesheetWidth ||
-    !body.spritesheetHeight ||
-    body.spritesheetWidth < MIN_SPRITE_DIM ||
-    body.spritesheetHeight < MIN_SPRITE_DIM
-  ) {
-    return {
-      ok: false,
-      status: 400,
-      error: "invalid_spritesheet",
-      message: `Spritesheet seems too small. Got ${body.spritesheetWidth}x${body.spritesheetHeight}, expected at least ${MIN_SPRITE_DIM}x${MIN_SPRITE_DIM} (ideal 1536x1872).`,
-      got: { width: body.spritesheetWidth, height: body.spritesheetHeight },
-    };
-  }
-  // Reject any URL outside the allowlist. Without this, a malicious
-  // submission could land javascript:, attacker.com, or LAN IPs into the
-  // pet detail page (XSS) and the install script (RCE on every viewer who
-  // pipes it through sh).
-  for (const field of ASSET_URL_FIELDS) {
-    if (!isAllowedAssetUrl(body[field])) {
-      return {
-        ok: false,
-        status: 400,
-        error: "invalid_asset_url",
-        field,
-        message: `${field} must be hosted on the petdex R2 bucket.`,
-      };
-    }
-  }
-  return null;
-}
-
-const ASSET_URL_FIELDS: ReadonlyArray<"zipUrl" | "spritesheetUrl" | "petJsonUrl"> =
-  ["zipUrl", "spritesheetUrl", "petJsonUrl"];
 
 /** Persist a submission. Caller is responsible for authn/ratelimit.
  *  Slug collisions get suffixed (boba -> boba-2) by resolveUniqueSlug.
@@ -150,9 +82,7 @@ export async function persistSubmission(
     // SMTP header injection defense — strip control chars from anything
     // that could end up in a header. The Resend SDK probably escapes, but
     // we don't trust user-controlled fields anywhere near a header.
-    const safeName = body.displayName
-      .replace(/[\r\n\t]+/g, " ")
-      .slice(0, 80);
+    const safeName = body.displayName.replace(/[\r\n\t]+/g, " ").slice(0, 80);
     void (async () => {
       try {
         const resend = new Resend(resendKey);
@@ -201,15 +131,6 @@ export function creditFromPrincipal(p: SubmissionPrincipal): {
     url: p.url ?? null,
     imageUrl: p.imageUrl ?? null,
   };
-}
-
-export function slugify(value: string): string {
-  return value
-    .toLowerCase()
-    .trim()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "")
-    .slice(0, 40);
 }
 
 export async function resolveUniqueSlug(base: string): Promise<string> {
